+++ platforms/aws/giantnetes/main.tf
@@ -125,6 +125,7 @@ locals {
     "CloudwatchForwarderEnabled"   = var.bastion_log_priority != "none" ? "true" : "false"
     "ClusterDomain"                = var.cluster_domain
     "ClusterName"                  = var.cluster_name
+    "DisableAPIFairness"           = var.disable_api_fairness
     "DockerCIDR"                   = var.docker_cidr
     "DockerRegistry"               = var.docker_registry
     "ETCDEndpoints"                = "https://etcd1.${var.base_domain}:2379,https://etcd2.${var.base_domain}:2379,https://etcd3.${var.base_domain}:2379"